Remove RoomStateService class by moving its static methods to the RoomService class
